# ms-userarea-ws/src/main/resources/openapi/openapi.yaml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: User Area API
  description: API to manage User Area
servers:
  - url: https://ua-local.dev/backend
    description: Local environment
  - url: https://euipo353.am.intrasoft-euipo.int/backend
    description: The dev environment
  - url: https://euipo345.am.intrasoft-euipo.int/backend
    description: The adaptive environment
  - url: https://ccvli-ecp613/backend
    description: The integration environment
  - url: https://ccvli-ecp623/backend
    description: The demo environment
  - url: https://ccvlt-ecp314/backend
    description: The test environment
paths:
...
  /correspondences/draft/{identifier}:
    delete:
      security:
        - bearerAuth: []
      tags:
        - Correspondence draft
      description: sets a draft to inactive
      summary: |
        Adds data to a created Draft
      operationId: deleteDraft
      parameters:
        - name: identifier
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: draft marked as inactive ok
        404:
          description: Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /correspondences/attachments/{draftId}:
    post:
      security:
        - bearerAuth: []
      tags:
        - Correspondence attachments
      summary: saves attachments for drafts
      operationId: createAttachment
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Draft Attachment data
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                    description: attachments
      responses:
        200:
          description: draft attachments uploaded ok
        401:
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
  /correspondences/attachments/{attachmentId}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Correspondence attachments
      summary: gets an attachment for the draft
      operationId: getAttachment
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        401:
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      security:
        - bearerAuth: []
      tags:
        - Correspondence attachments
      summary: deletes an attachment for the draft
      operationId: deleteAttachment
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        401:
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
  /correspondences/attachments/draft/{draftAttachmentId}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Correspondence attachments
      summary: gets an attachment for the draft reply
      operationId: draftAttachmentId
      parameters:
        - name: draftAttachmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        401:
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /signatures/modify/application/{applicationType}/{applicationNumber}:
    put:
      security:
        - bearerAuth: []
      operationId: modifyApplication
      tags:
        - Signature Modify Application
      summary: Modify application from signature tab
      parameters:
        - name: applicationType
          in: path
          required: true
          schema:
            type: string
        - name: applicationNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        404:
          description: No application record found
          content: { }
...
components:
  securitySchemes:
    bearerAuth:
      type: apiKey # arbitrary name for the security scheme
      in: cookie
      name: JSESSIONID_UAB
  schemas:
    ApplicationResponse:
      type: object
      description: the application response
      properties:
        applicationNumber:
          type: string
        resumeUrl:
          type: string
...
    Error:
      type: object
      properties:
        title:
          type: string
          description: |
            A short, human-readable summary of the problem type
        detail:
          type: string
          description: |
            A human-readable explanation specific to this occurrence of the problem
        status:
          type: integer
          format: int32
          description: |
            The HTTP status code generated by the origin server for this occurrence of the problem
        type:
          type: string
          format: uri
          description: |
            A URI reference [RFC3986] that identifies the
                  problem type.  This specification encourages that, when
                  dereferenced, it provide human-readable documentation for the
                  problem type
        instance:
          type: string
          format: uri-reference
          description: |
            A URI reference that identifies the specific occurrence of the problem.
            It may or may not yield further information if dereferenced
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
      required:
        - title
        - detail
...